+++ Pandora/Paradigm/Basis/Identity.hs
@@ -18,6 +18,7 @@ import Pandora.Pattern.Object.Chain (Chain ((<=>)))
 import Pandora.Pattern.Object.Semigroup (Semigroup ((+)))
 import Pandora.Pattern.Object.Monoid (Monoid (zero))
 import Pandora.Pattern.Object.Ringoid (Ringoid ((*)))
+import Pandora.Pattern.Object.Quasiring (Quasiring (one))
 import Pandora.Pattern.Object.Semilattice (Infimum ((/\)), Supremum ((\/)))
 import Pandora.Pattern.Object.Lattice (Lattice)
 import Pandora.Pattern.Object.Group (Group (inverse))
@@ -70,6 +71,9 @@ instance Monoid a => Monoid (Identity a) where
 
 instance Ringoid a => Ringoid (Identity a) where
 	Identity x * Identity y = Identity $ x * y
+
+instance Quasiring a => Quasiring (Identity a) where
+	 one = Identity one
 
 instance Infimum a => Infimum (Identity a) where
 	Identity x /\ Identity y = Identity $ x /\ y